+++ core-ng/src/test/java/core/framework/impl/web/service/WebServiceInterfaceValidatorTest.java
@@ -2,8 +2,11 @@ package core.framework.impl.web.service;
 
 import core.framework.impl.web.bean.RequestBeanMapper;
 import core.framework.impl.web.bean.ResponseBeanTypeValidator;
+import core.framework.util.Types;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import java.lang.reflect.Method;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -24,8 +27,21 @@ class WebServiceInterfaceValidatorTest {
     }
 
     @Test
-    void validateRequestBeanType() {
-        assertThatThrownBy(() -> validator.validateRequestBeanType(Integer.class, TestWebService.class.getDeclaredMethod("get", Integer.class)))
+    void validateRequestBeanType() throws NoSuchMethodException {
+        Method method = TestWebService.class.getDeclaredMethod("get", Integer.class);
+
+        assertThatThrownBy(() -> validator.validateRequestBeanType(Integer.class, method))
                 .isInstanceOf(Error.class).hasMessageContaining("if it is path param, please add @PathParam");
+    }
+
+    @Test
+    void validateResponseBeanType() throws NoSuchMethodException {
+        Method method = TestWebService.class.getDeclaredMethod("get", Integer.class);
+
+        assertThatThrownBy(() -> validator.validateResponseBeanType(Integer.class, method))
+                .isInstanceOf(Error.class).hasMessageContaining("response bean type must be bean class");
+
+        assertThatThrownBy(() -> validator.validateResponseBeanType(Types.map(String.class, String.class), method))
+                .isInstanceOf(Error.class).hasMessageContaining("response bean type must be bean class");
     }
 }